requirements.txt
seaborn
transformers
torchvision
ftfy
regex
tqdm
adal==1.2.7
aiofiles==22.1.0
aiohttp==3.8.3
aiokafka==0.8.0
aiosignal==1.3.1
alembic==1.8.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argcomplete==2.0.0
async-timeout==4.0.2
attrs==22.1.0
azure-common==1.1.28
azure-core==1.26.1
azure-graphrbac==0.61.1
azure-mgmt-authorization==3.0.0
azure-mgmt-containerregistry==10.0.0
azure-mgmt-core==1.3.2
azure-mgmt-keyvault==10.1.0
azure-mgmt-resource==21.2.1
azure-mgmt-storage==20.1.0
azureml-core==1.48.0
backports-tempfile==1.0
backports-weakref==1.0.post1
bcrypt==4.0.1
boto3==1.26.31
botocore==1.29.31
brotli==1.0.9
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorlog==6.7.0
comet-ml==3.31.20
commonmark==0.9.1
configobj==5.0.6
contextlib2==21.6.0
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
databricks-cli==0.17.4
distlib==0.3.6
docker==6.0.1
docker-pycreds==0.4.0
dulwich==0.20.50
entrypoints==0.4
everett==3.1.0
fastapi==0.88.0
filelock==3.8.2
flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
gevent==22.10.2
geventhttpclient==2.0.2
gitdb==4.0.10
gitpython==3.1.29
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
grpcio==1.51.1
gunicorn==20.1.0
h11==0.14.0
humanfriendly==10.0
hydra-colorlog==1.2.0
hydra-core==1.2.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.8.0
jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==2.2.0
jsonschema==4.17.3
kafka-python==2.0.2
kiwisolver==1.4.4
knack==0.10.1
kubernetes==25.3.0
lightning-utilities==0.4.2
llvmlite==0.39.1
mako==1.2.4
markdown==3.4.1
markupsafe==2.1.1
matplotlib==3.6.2
mlflow==2.0.1
mlserver==1.2.1
mlserver-mlflow==1.2.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.3
ndg-httpsclient==0.5.1
nltk==3.8
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
orjson==3.8.3
packaging==21.3
pandas==1.5.2
paramiko==2.12.0
pathspec==0.10.3
pathtools==0.1.2
pillow==9.3.0
pkginfo==1.9.2
pkgutil-resolve-name==1.3.10
platformdirs==2.6.0
portalocker==2.6.0
prometheus-client==0.15.0
prometheus-flask-exporter==0.21.0
promise==2.3
protobuf
psutil==5.9.4
py-grpc-prometheus==0.7.0
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pygments==2.13.0
pyjwt==2.6.0
pynacl==1.5.0
pyopenssl==22.1.0
pyparsing==3.0.9
pyrootutils==1.0.4
pyrsistent==0.19.2
pysftp==0.2.9
pysocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-rapidjson==1.9
pytorch-lightning==1.8.4.post0
pytz==2022.6
pyyaml==6.0
querystring-parser==1.2.4
randaugment==1.0.0
regex==2022.10.31
requests==2.28.1
requests-auth-aws-sigv4==0.7
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==12.6.0
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.0
scipy==1.9.3
secretstorage==3.3.3
semantic-version==2.10.0
sentry-sdk==1.11.1
setproctitle==1.3.2
shap==0.41.0
shortuuid==1.0.11
simplejson==3.18.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
sqlalchemy==1.4.45
sqlparse==0.4.3
starlette==0.22.0
starlette-exporter==0.14.0
tabulate==0.9.0
tensorboardx==2.5.1
termcolor==2.1.1
threadpoolctl==3.1.0
torch==1.13.0
torchmetrics==0.11.0
tqdm==4.64.1
tritonclient==2.28.0
typing-extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.17.1
wandb==0.13.7
wcwidth==0.2.5
websocket-client==1.3.3
werkzeug==2.2.2
wrapt==1.14.1
wurlitzer==3.0.3
yarl==1.8.2
zipp==3.11.0
zope-event==4.6
zope-interface==5.5.2
